Drop the casts decoded DBC signals no longer need
With the generator handing out each signal's narrowest exact type, most
consumer casts were converting a value to the type it already had, or
widening an 8 or 16 bit integer into a float, which is lossless.

  motec_m1        201 casts removed; 1 static_cast left in the file, and it
                  is not on a decoded signal
  motec_pdm       130 casts and 23 `!= 0` removed from the fill tables;
                  lambdas over integer signals declare `-> float`
  megasquirt      20 casts removed
  motec_ltc       8 removed; the 7 static_cast<bool> on the fault flags stay,
                  because those 1-bit signals carry value tables and are enums
  msel            enum bytes via std::to_underlying; frame members and
                  tenthsToMilliseconds are float, so msel_master_relay drops
                  its 4 casts too

MotecM1Temperatures' five temperatures go from Int8 to Int16. The fields are
8 bit with a -40 offset, so they read -40..215 C, and everything over 127 C
wrapped negative. Recordings made before this change will misread those five
fields.

megasquirt_test_frames computes in float, and fixes AFR1 and afrtgt1: they
were assigned 140 and 145 into scale-0.1 fields, which saturated to 25.5.

Verified on the bus: megasquirt_test_frames through the megasquirt node
publishes AFR 14.1 against a 14.5 target; an M1 0x649 frame with coolant raw
200 and oil raw 215 publishes 160 and 175 C, and raw 0 reads -40. Whole tree
builds under -Werror; 148 unit and 10 gui tests pass. A one-off
-Wconversion -Wfloat-conversion pass over every consumer finds no narrowing
of a decoded signal beyond E888's two int32_t temperatures (-200..65335,
exact in a float) and the mock's mux index.

+++ libs/msel/include/msel/protocol.h
@@ -216,9 +216,9 @@ uint16_t encodeConfigWord(const Config& config);
 // what actually arrived instead of "unknown".
 struct StatusFrame
 {
-    double voltageOut { 0.0 };           // V, on the output terminal
-    double loadCurrent { 0.0 };          // A, positive when discharging
-    double temperatureInternal { 0.0 };  // degrees C
+    float voltageOut { 0.0f };           // V, on the output terminal
+    float loadCurrent { 0.0f };          // A, positive when discharging
+    float temperatureInternal { 0.0f };  // degrees C
     uint8_t warnings { 0u };             // bitmask of Warning
     uint8_t statusRaw { 0u };
     std::optional<Status> status;
@@ -226,7 +226,7 @@ struct StatusFrame
 
 struct InfoFrame
 {
-    double voltageIn { 0.0 };            // V, on the input (battery) terminal
+    float voltageIn { 0.0f };            // V, on the input (battery) terminal
     uint16_t serialNo { 0u };
     Config config;
     std::chrono::milliseconds timeSinceShutdown { 0 };
