+++ nodes/megasquirt/main.cpp
@@ -30,34 +30,34 @@ static void publish_dash(const dbc_megasquirt_dash_data_parser::db_t& db, pub_su
     const auto& dash4 = db.megasquirt_dash4;
 
     // Frame 0
-    out.setRpm(static_cast<uint16_t>(dash0.rpm));
-    out.setMapKpa(static_cast<float>(dash0.map));
-    out.setTpsPct(static_cast<float>(dash0.tps));
-    out.setCoolantTempF(static_cast<float>(dash0.clt));
+    out.setRpm(dash0.rpm);
+    out.setMapKpa(dash0.map);
+    out.setTpsPct(dash0.tps);
+    out.setCoolantTempF(dash0.clt);
 
     // Frame 1
-    out.setIgnitionAdvanceDeg(static_cast<float>(dash1.adv_deg));
-    out.setIntakeAirTempF(static_cast<float>(dash1.mat));
-    out.setInjPw1Ms(static_cast<float>(dash1.pw1));
-    out.setInjPw2Ms(static_cast<float>(dash1.pw2));
+    out.setIgnitionAdvanceDeg(dash1.adv_deg);
+    out.setIntakeAirTempF(dash1.mat);
+    out.setInjPw1Ms(dash1.pw1);
+    out.setInjPw2Ms(dash1.pw2);
 
     // Frame 2
-    out.setSeqPw1Ms(static_cast<float>(dash2.pwseq1));
-    out.setEgt1F(static_cast<float>(dash2.egt1));
-    out.setEgoCorrectionPct(static_cast<float>(dash2.egocor1));
-    out.setAfr1(static_cast<float>(dash2.AFR1));
-    out.setAfrTarget1(static_cast<float>(dash2.afrtgt1));
+    out.setSeqPw1Ms(dash2.pwseq1);
+    out.setEgt1F(dash2.egt1);
+    out.setEgoCorrectionPct(dash2.egocor1);
+    out.setAfr1(dash2.AFR1);
+    out.setAfrTarget1(dash2.afrtgt1);
 
     // Frame 3
-    out.setKnockRetardDeg(static_cast<float>(dash3.knk_rtd));
-    out.setSensor1(static_cast<float>(dash3.sensors1));
-    out.setSensor2(static_cast<float>(dash3.sensors2));
-    out.setBatteryVolts(static_cast<float>(dash3.batt));
+    out.setKnockRetardDeg(dash3.knk_rtd);
+    out.setSensor1(dash3.sensors1);
+    out.setSensor2(dash3.sensors2);
+    out.setBatteryVolts(dash3.batt);
 
     // Frame 4
-    out.setLaunchTimingDeg(static_cast<float>(dash4.launch_timing));
-    out.setTcRetard(static_cast<float>(dash4.tc_retard));
-    out.setVssMps(static_cast<float>(dash4.VSS1));
+    out.setLaunchTimingDeg(dash4.launch_timing);
+    out.setTcRetard(dash4.tc_retard);
+    out.setVssMps(dash4.VSS1);
 
     pub.put();
 }
